services/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
python-dotenv==1.0.0
